refactor: split into format_table* files
