+++ src/format_table.h
@@ -1,0 +1,8 @@
+#ifndef FORMAT_TABLE_H
+#define FORMAT_TABLE_H
+
+#include "jv.h"
+
+jv format_table(jv input);
+
+#endif // FORMAT_TABLE_H
