--- src/table_format.h
@@ -1,9 +1,0 @@
-#ifndef TABLE_FORMAT_H
-#define TABLE_FORMAT_H
-
-#include "jq.h"
-#include "jv.h"
-
-jv format_table(jv input);
-
-#endif // TABLE_FORMAT_H
